+++ lib/src/extension/navigator_ext.dart
@@ -3,27 +3,31 @@ import 'package:go_router/go_router.dart';
 import '../manager/app_manager.dart';
 import '../utils/print.dart';
 
+const String _slash = '/';
+
 extension NavigatorExt on List<String> {
+  /// If [AppManager().slashResolution] is defined, it will be used when / is
+  /// received.
   void route({
-    Map<String, dynamic>? query,
+    Map<String, dynamic> query = const {},
   }) {
-    String slash = '/';
-    if (this[0] == '') {
-      removeAt(0);
+    final manager = AppManager();
+
+    removeWhere((str) => str.isEmpty);
+
+    final clean = map((i) => i.replaceAll(_slash, '')).toList();
+
+    var path = '$_slash${clean.join(_slash)}';
+
+    if (manager.slashResolution.isNotEmpty && path == _slash) {
+      path = '$_slash${manager.slashResolution.replaceAll(_slash, '')}';
     }
 
-    final clean = <String>[];
-    for (var i in this) {
-      var a = i.replaceAll('/', '');
-      clean.add(a);
+    if (manager.debug) {
+      printDebug('Routing to path: $path');
     }
-    var path = '$slash${clean.join(slash)}';
-    printDebug('#### going to path: $path');
 
-    if (path == '/') {
-      path = '/chats';
-    }
-    AppManager().lastExtra.addAll(query ?? {});
-    AppManager().context.go(path);
+    manager.lastExtra.addAll(query);
+    manager.context.go(path);
   }
 }