+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Renamed `AppManager().withRouterAndModel` to `AppManager().configure`.
+* Added `debug`, `lowPerformance`, `refreshLatency`, `slashResolution` and `screenVisible` params to `AppManager().configure`.
+
 ## 1.0.1
 
 * Bump `object` version.